refactor: Add suffix "_model" to course, course_file, evaluation and lecture model files
