+++ server/app/models/__init__.py
@@ -1,4 +1,4 @@
-from .course import Course
-from .course_file import CourseFile
-from .evaluation import Evaluation
-from .lecture import Lecture
+from .course_model import Course
+from .course_file_model import CourseFile
+from .evaluation_model import Evaluation
+from .lecture_model import Lecture
