+++ server/app/models/course_model.py
