+++ server/app/models/lecture_model.py
